ADD file structure modifications
